"Update styling and images in various components; change color scheme and add plan images"
This commit includes updates to the styling and images in several components. The color scheme has been changed from the previous orange to a new accent color. Additionally, plan images have been added to the Plans component. The specific changes include:

* Updating the background image and removing the hardcoded height in Home component
* Changing the color scheme in Header component
* Adjusting the layout and styling in Gallery component
* Adding plan images in Plans component

Overall, these changes improve the visual appearance and consistency of the application.