"Update components with new styling and images; add plan pictures"
This commit includes updates to the Header.js, Home.js, and Plan.js components. In Header.js, the color of the "Genesis" text in the logo has been changed from #f35903 to accent, and some minor changes have been made to the navigation menu for mobile devices. In Home.js, the background image has been removed and the TextContainer class has been added to the text. In Plan.js, an image has been added to each plan and the layout of the plans has been changed to a grid. The grid layout is responsive and changes based on the screen size.